+++ LibOS/shim/src/bookkeep/shim_handle.c
@@ -72,7 +72,7 @@ static inline int init_tty_handle (struct shim_handle * hdl, bool write)
     hdl->dentry = dent;
     hdl->flags = O_RDWR|O_APPEND|0100000;
 
-    int size;
+    size_t size;
     char * path = dentry_get_path(dent, true, &size);
     if (path)
         qstrsetstr(&hdl->path, path, size);
@@ -102,7 +102,7 @@ static inline int init_exec_handle (struct shim_thread * thread)
                       &exec->dentry, fs);
         set_handle_fs(exec, fs);
         if (exec->dentry) {
-            int len;
+            size_t len;
             const char * path = dentry_get_path(exec->dentry, true, &len);
             qstrsetstr(&exec->path, path, len);
         }
